CNDE-2923: error mapping and unit tests
